mysite/requirements.txt
asgiref==3.5.0
Django==4.0.2
django-froala-editor==4.0.13
Pillow==9.0.1
sqlparse==0.4.2
tzdata==2021.5
